% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_control_metrics.R
\name{bathresh}
\alias{bathresh}
\title{Get BeadArray control outcomes from a matrix of metric signals}
\usage{
bathresh()
}
\value{
Data frame of minimum BeadArray quality thresholds.
}
\description{
Get Illumina's prescribed minimum quality thresholds for BeadArray metrics.
}
\examples{
dfthresh <- bathresh()
}
\references{
1. Illumina, “Illumina Genome Studio Methylation Module v1.8,” Nov. 2010.

2. Illumina, “BeadArray Controls Reporter Software Guide,” Oct. 2015.
}
\seealso{
bactrl
}
