% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{data.mdpost}
\alias{data.mdpost}
\title{Retrieve available sample metadata}
\usage{
data.mdpost(dbn = "remethdb2.h5", dsn = "mdpost")
}
\arguments{
\item{dbn}{Path to HDF5 database file.}

\item{dsn}{Name or group path to HDF5 dataset containing 
the sample metadata and learned annotations.}
}
\value{
A `data.frame` of all the sample metadata
}
\description{
Retrieves all sample metadata from an HDF5 database.
}
\examples{
# get all available sample metadata
mdp <- data.mdpost(dbn = "remethdb2.h5", dsn = "mdpost")
}
\seealso{
hread()
}
