% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{data.nb}
\alias{data.nb}
\title{Get noob-normalized Beta-values.}
\usage{
data.nb(
  dbn = "remethdb.h5",
  dsn = "noobbeta",
  which.gsm = sample(35700, 5),
  which.cg = c(1:485512)
)
}
\arguments{
\item{dbn}{Name of HDF5 database ('.h5') file.}

\item{dsn}{Name of HDF5 dataset containing noob-normalized Beta-values.}

\item{which.gsm}{Numeric indices of samples whose data to retrieve.}

\item{which.cg}{Numeric indices of CpG probes whose data to retrieve.}
}
\value{
Returns subset of noob-normalized Beta-values HDF5 dataset matching query.
}
\description{
Retrieves noob-normalized Beta-values dataset from HDF5 database file.
}
