% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{data.rgsignal}
\alias{data.rgsignal}
\title{Get raw probe signal plots}
\usage{
data.rgsignal(
  dbn = "remethdb.h5",
  dsn = c("redsignal", "greensignal"),
  which.gsm = sample(35700, 5),
  which.cg = c(1:622399)
)
}
\arguments{
\item{dbn}{Name of HDF5 database ('.h5') file.}

\item{dsn}{Name or group path to a raw signal dataset (either 'redsignal' or 'greensignal').}

\item{which.gsm}{Numeric indices of samples (signal dataset rows) whose data to retrieve.}

\item{which.cg}{Numeric indices of CpG probes (signal dataset columns) whose data to retrieve.}
}
\value{
Returns the subset of the raw signal dataset matching the query.
}
\description{
Retrieves raw signal from either red or green array color channel HDF5 dataset.
}
