% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_db.R
\name{dldb}
\alias{dldb}
\alias{dldb_h5se_gr}
\alias{dldb_h5se_gm}
\alias{dldb_h5se_rg}
\alias{dldb_h5_rg}
\alias{dldb_h5se_test}
\alias{dldb_h5_test}
\title{Functions to download latest or recommended versions of the DNAm datasets.}
\usage{
dldb_h5se_gr(verbose = FALSE, dfp = "downloads")

dldb_h5se_gm(verbose = FALSE, dfp = "downloads")

dldb_h5se_rg(verbose = FALSE, dfp = "downloads")

dldb_h5_rg(verbose = FALSE, dfp = "downloads")

dldb_h5se_test(verbose = FALSE, dfp = "downloads")

dldb_h5_test(verbose = FALSE, dfp = "downloads")
}
\arguments{
\item{verbose}{Whether to return verbose status messages (default FALSE)}

\item{dfp}{Download destination directory}
}
\value{
file path
}
\description{
Functions that download the latest or recommended versions of DNAm datasets by calling get_rmdl().
 Data files can be of either `HDF5-SummarizedExperiment` or `HDF5` type. Data types include red and green 
 signals for HDF5 files, or `RGChannelSet` ("rg"), `MethylSet` ("gm"), or `GenomicRatioSet` ("gr") for 
 HDF5-SummarizedExperiment files. Files include sample metadata.
}
\examples{
# download_h5se_gr()
# download_h5se_gm()
# download_h5se_rg()
# download_h5_rg()
download_h5se_test()
download_h5_test()
}
\seealso{
get_rmdl, getrg, data_mdpost
}
