% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_rmdl}
\alias{get_rmdl}
\title{Get DNAm assay data.}
\usage{
get_rmdl(
  fn = NULL,
  show.files = FALSE,
  which.class = c("rg", "gm", "gr", "test"),
  which.type = c("h5se", "h5"),
  url = "https://recount.bio/data/",
  dfp = "downloads",
  download = TRUE,
  verbose = TRUE,
  sslver = FALSE
)
}
\arguments{
\item{which.class}{Class of file to download (either `rg`, `gm`, `gr`, or `test`).}

\item{which.type}{Type of file (either `h5` for HDF5 file or `h5se` for HDF5-SummarizedExperiment).}

\item{url}{Server URL containing assay data.}

\item{dfp}{Target local directory for downloaded files (default "downloads").}

\item{download}{Whether to download (TRUE) or return queried filename (FALSE).}

\item{verbose}{Whether to return verbose messages.}

\item{sslver}{Whether to use server certificate check (default FALSE).}
}
\value{
New filepath to dir with downloaded data.
}
\description{
Uses RCurl to recursively download latest H5SE and HDF5 data objects the from server.
}
\examples{
get_rmdl("h5se-test_gr", verbose = TRUE)
}
