% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_db.R
\name{getdb}
\alias{getdb}
\alias{getdb_h5se_test}
\alias{getdb_h5_test}
\title{Functions to download and access database files.}
\usage{
getdb_h5se_test(name = NULL, dfp = "downloads")

getdb_h5_test(name = NULL, dfp = "downloads")
}
\arguments{
\item{name}{Name of database file, optional (default NULL).}

\item{dfp}{Folder to search for object, optional (default "downloads").}
}
\value{
SummarizedExperiment object (for h5se) or file path for (h5).
}
\description{
Combines download and load for database files. Default behavior downloads database files to 
"downloads" directory.
}
\examples{
getdb_h5se_test()
getdb_h5se_test()
}
\seealso{
dldb, get_rmdl
}
