% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getgs}
\alias{getgs}
\title{Get noob-normalized Beta-value data as either a list of `data.frame` or an `GenomicRatioSet` object}
\usage{
getgs(
  dbn,
  gsmv,
  cgv,
  dat.type = c("se", "df"),
  dsv = c("noobbeta"),
  metadata = TRUE
)
}
\arguments{
\item{dbn}{Name of the HDF5 database file.}

\item{gsmv}{Vector valid GSM IDs (rows) to query in the raw signal datasets.}

\item{cgv}{Vector of valid CpG probe addresses (columns) to query in the raw signal datasets.}

\item{dsv}{Vector of raw signal datasets or group paths to query, including both the red channel 'redsignal' and green channel 'greensignal' datasets.}

\item{metadata}{Whether to access available postprocessed metadata for queries samples.}

\item{data.type}{Format for returned query matches, either as datasets 'df' or `GenomicRatioSet` 'se' object.}
}
\value{
Returns either an `GenomicRatioSet` or list of `data.frame` objects from dataset query matches.
}
\description{
Retrieves query matches from raw signal HDF5 datasets. Handles identity queries to rows (GSM IDs) or columns (CpG probe addresses). Returns query matches either as a list of 2 `data.frame`s or a signle `RGChannelSet` object.
}
