% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{grse}
\alias{grse}
\title{Form an object of class `GenomicRatioSet` from a raw signal dataset query}
\usage{
grse(ldat)
}
\arguments{
\item{ldat}{List of raw signal data query results. Must include 2 `data.frame` objects named 'redsignal' and 'greensignal'.}
}
\value{
Returns a `GenomicRatioSet` object from raw signal dataset queries.
}
\description{
Forms a `GenomicRatioSet` object from an HDF5 databse file query to the red and green raw signal tables. See `getgr()` function for implementation.
}
