% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{hdcon}
\alias{hdcon}
\title{Connect to an HDF5 dataset file.}
\usage{
hdcon(dsn = "noobbeta", hfile = hfcon())
}
\arguments{
\item{dsn}{Name of dataset or group of dataset to connect with.}

\item{hfile}{Connection to an HDF5 database file.}
}
\value{
HDF5 database connection object.
}
\description{
Get a dataset connection object from an HDF5 database ('.h5') file.
}
