% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{hfcon}
\alias{hfcon}
\title{Connection to HDF5 database file}
\usage{
hfcon(dbn.path = "remethdb.h5")
}
\arguments{
\item{dbn.path}{Path to HDF5 database file.}
}
\value{
An HDF5 file connection object.
}
\description{
Get '.h5' file connection (db connection).
}
