% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{matchds.1to2}
\alias{matchds.1to2}
\title{Match two datasets}
\usage{
matchds.1to2(
  ds1,
  ds2,
  mi1 = c("rows", "columns"),
  mi2 = c("rows", "columns"),
  subset.match = FALSE
)
}
\arguments{
\item{ds1}{First dataset to match}

\item{ds2}{Second dataset to match}

\item{mi1}{Match index of ds1 (either "rows" or "columns")}

\item{mi2}{Match index of ds2 (either "rows" or "columns")}

\item{subset.match}{If index lengths don't match, match on the 
common subset instead}
}
\value{
A list of the matched datasets.
}
\description{
Match the character vectors of row or column names 
for 2 datasets.
}
