% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnam_search_index.R
\name{query_si}
\alias{query_si}
\title{query_si}
\usage{
query_si(sample_idv, fh_csv_fname, index_dict_fname, lkval = c(1, 2))
}
\arguments{
\item{lkval}{Vector of K nearest neighbors to return per query (optional, 
int, c(1,2)).}

\item{query_data}{Vector of feature hashed data comprising the query with 
same length (for a vector) or ncol (for a table) as in the queried search 
index data (required, numeric vector, matrix, or data frame).}

\item{si_fname}{Name/path of queried search index (required, string).}
}
\value{

}
\description{
Query an HNSW search index. Does K Nearest Neighbors lookup on a previously 
saved search index object, returning the K nearest neighbors of the queried 
sample(s). The `query_si()` function returns verbose output, which can be 
silenced with suppressMessages()`.
}
