% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnam_search_index.R
\name{setup_sienv}
\alias{setup_sienv}
\title{setup_sienv}
\usage{
setup_sienv(
  env.name = "dnam_si_hnswlib",
  pkgv = c("hnswlib==0.5.1", "pandas==1.2.2", "numpy==1.20.1", "mmh3==3.0.0",
    "h5py==3.2.1")
)
}
\arguments{
\item{env.name}{Name of the new virtual environment (required, 
"dnam_si-hnswlib")}

\item{pkgv}{Vector of the dependencies and their versions for the new
virtual environment (required, format: "packagename==versionnum").}
}
\value{
New basilisk environment object.
}
\description{
Set up a new virtual environment for search index construction using 
the basilisk package.
}
