% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{data.mdpost}
\alias{data.mdpost}
\title{Retrieve samples metadata.}
\usage{
data.mdpost(dbn.path = "remethdb.h5", dsn = "mdpost")
}
\arguments{
\item{dbn.path}{Path to HDF5 database file.}

\item{dsn}{Name or group path to HDF5 dataset containing postprocessed metadata.}
}
\value{
Postprocessed metadata as a `data.frame`.
}
\description{
Retrieves sample postprocessed metadata from an HDF5 database.
}
