% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{data_mdpost}
\alias{data_mdpost}
\title{Retrieve HDF5 sample metadata}
\usage{
data_mdpost(dbn = "remethdb2.h5", dsn = "mdpost")
}
\arguments{
\item{dbn}{Path to HDF5 database file.}

\item{dsn}{Name or group path to HDF5 dataset containing sample metadata.}
}
\value{
A `data.frame` of all the sample metadata
}
\description{
Retrieves sample metadata from HDF5 database.
}
\examples{
# notrun:
# mdp <- data_mdpost(dbn = "remethdb2.h5", dsn = "mdpost")
}
