% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{data_mdpost}
\alias{data_mdpost}
\title{Retrieve available sample metadata}
\usage{
data_mdpost(dbn = "remethdb2.h5", dsn = "mdpost")
}
\arguments{
\item{dbn}{Path to HDF5 database file.}

\item{dsn}{Name or group path to HDF5 dataset containing 
the sample metadata and learned annotations.}
}
\value{
A `data.frame` of all the sample metadata
}
\description{
Retrieves all sample metadata from an HDF5 database.
}
\examples{
h5test.name <- "remethdb-h5_rg-test_0-0-1_1590090412.h5"
path <- system.file("extdata", h5test.name, package = "recountmethylation")
fn = list.files(path)
dbpath = paste0(path, "/", fn)
mdp <- data_mdpost(dbn = dbpath, dsn = "mdpost")
dim(mdp) # [1]  2 19
}
\seealso{
hread()
}
