% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{data_mdpost}
\alias{data_mdpost}
\title{Query and store sample metadata.}
\usage{
data_mdpost(dbn = "remethdb2.h5", dsn = "mdpost")
}
\arguments{
\item{dbn}{Path to HDF5 database file.}

\item{dsn}{Name or group path to HDF5 dataset containing postprocessed metadata.}
}
\value{
Postprocessed metadata as a `data.frame`.
}
\description{
Retrieves sample postprocessed metadata from an HDF5 database.
}
\examples{
# get all available sample metadata
mdp <- data_mdpost(dbn = "remethdb2.h5", dsn = "mdpost")
}
\seealso{
hread()
}
