% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{gds_idat2rg}
\alias{gds_idat2rg}
\title{Handles IDAT queries to GDS and returns assay data as `RGChannelSet`.}
\usage{
gds_idat2rg(
  gsmvi,
  rmdl = TRUE,
  ext = "gz",
  dfp = "./idats/",
  burl = paste0("ftp://ftp.ncbi.nlm.nih.gov/", "geo/samples/")
)
}
\arguments{
\item{gsmvi}{A vector of GSM IDs (alphanumeric character strings).}

\item{rmdl}{Whether to remove downloaded .*idat files (default TRUE).}

\item{ext}{Filename extension (default "gz").}

\item{dfp}{Destination file dir for idats.}

\item{burl}{Base URL string for idat query (default "ftp://ftp.ncbi.nlm.nih.gov/geo/samples/").}
}
\value{
object of class `RGChannelSet`
}
\description{
Queries and downloads GSM IDAT files in GDS, and returns
assay data as an `RGChannelSet`.
('.h5') file and return the indexed table subset.
}
\examples{
gsmvi <- c("GSM2465267", "GSM2814572")
rg <- gds_idat2rg(gsmvi)
}
