% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{gds_idatquery}
\alias{gds_idatquery}
\title{IDATs query}
\usage{
gds_idatquery(
  gsmvi,
  ext = "gz",
  expand = TRUE,
  sys.cmd = "gunzip ",
  verbose = FALSE,
  dfp = "./idats/",
  burl = paste0("ftp://ftp.ncbi.nlm.nih.gov/", "geo/samples/")
)
}
\arguments{
\item{gsmvi}{Param description}

\item{ext}{Filename extension}

\item{expand}{Whether to expand compressed files.}

\item{sys.cmd}{System command to expand compressed files (if expand TRUE).}

\item{verbose}{Whether to show verbose messages (TRUE/FALSE)}

\item{dfp}{Download directory}

\item{burl}{Base URL string for RCurl query}
}
\value{
Describes returned object
}
\description{
Queries GDS IDATs and downloads, with exception handling.
Files are downloaded
}
\examples{
# Not run
# gsmvi <- c("GSM2465267", "GSM2814572")
# gds_idatquery(gsmvi)
}
