% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get.rmdl}
\alias{get.rmdl}
\title{Get DNAm assay data.}
\usage{
get.rmdl(
  which.dn = c("h5se-test_gr", "h5se_gr", "h5se_gm", "h5se_rg", "\\\\.h5"),
  url = "https://recount.bio/data/",
  dfp = "data",
  verbose = TRUE
)
}
\arguments{
\item{url}{Server URL containing assay data.}

\item{dfp}{Data file path. Use "" for working dir.}

\item{verbose}{Whether to return verbose messages.}

\item{which.file}{Type of data object to be downloaded.}
}
\value{
New filepath to dir with downloaded data.
}
\description{
Uses RCurl to recursively download latest H5SE and HDF5 data objects the from server.
}
\examples{
get_rmdl("h5se-test_gr", verbose = TRUE)
}
