% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_db.R
\name{get_rmdl}
\alias{get_rmdl}
\title{Get DNAm assay data.}
\usage{
get_rmdl(
  which.class = c("rg", "gm", "gr", "test"),
  which.type = c("h5se", "h5"),
  fn = NULL,
  dfp = "downloads",
  url = "https://recount.bio/data/",
  show.files = FALSE,
  download = TRUE,
  verbose = TRUE,
  sslver = FALSE
)
}
\arguments{
\item{which.class}{Either "rg", "gm", "gr", or "test" for RGChannelSet, 
MethylSet, GenomicRatioSet, or 2-sample subset (default "test").}

\item{which.type}{Either "h5se" for an HDF5-SummarizedExperiment (default) 
or "h5" for an HDF5 database.}

\item{fn}{Name of file on server to download (optional, default NULL).}

\item{dfp}{Download destination directory (default "downloads").}

\item{url}{The server URL to locate files for download.}

\item{show.files}{Whether to print server file data to console (default 
FALSE).}

\item{download}{Whether to download (TRUE) or return queried filename 
(FALSE).}

\item{verbose}{Whether to return verbose messages.}

\item{sslver}{Whether to use server certificate check (default FALSE).}
}
\value{
New filepath to dir with downloaded data.
}
\description{
Uses RCurl to recursively download latest H5SE and HDF5 data objects from 
the server. This is currently wrapped in the getdb() functions.
}
\seealso{
servermatrix(), getURL(), loadHDF5SummarizedExperiment(), h5ls()
}
