% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_rmdl}
\alias{get_rmdl}
\title{Get DNAm assay data.}
\usage{
get_rmdl(
  which.dn = c("h5se-test_gr", "h5se_gr", "h5se_gm", "h5se_rg", "\\\\.h5"),
  url = "https://recount.bio/data/",
  dfp = "",
  download = c("download", "filename"),
  verbose = TRUE,
  sslver = FALSE
)
}
\arguments{
\item{which.dn}{Type of data dir to be downloaded.}

\item{url}{Server URL containing assay data.}

\item{dfp}{Target local directory for downloaded files.}

\item{download}{Whether to download or just return latest filename.}

\item{verbose}{Whether to return verbose messages.}

\item{sslver}{Whether to use server certificate check (default FALSE).}
}
\value{
New filepath to dir with downloaded data.
}
\description{
Uses RCurl to recursively download latest H5SE and HDF5 data objects the from server.
}
\examples{
get_rmdl("h5se-test_gr", verbose = TRUE)
}
