% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_db.R
\name{get_rmdl}
\alias{get_rmdl}
\title{Get DNAm assay data.}
\usage{
get_rmdl(
  which.class = c("rg", "gm", "gr", "test"),
  which.type = c("h5se", "h5"),
  fn = NULL,
  dfp = "downloads",
  url = "https://recount.bio/data/",
  show.files = FALSE,
  download = TRUE,
  tryload = TRUE,
  verbose = TRUE,
  sslver = FALSE
)
}
\arguments{
\item{which.class}{Class of file to download (either "rg" for `RGChannelSet`, 
"gm" for `MethylSet`, "gr" for `GenomicRatioSet`, or "test" for the test dataset).}

\item{which.type}{Type of file. Either "h5se" for HDF5-SummarizedExperiment (default) or 
`h5` for HDF5 database.}

\item{fn}{Name of server file to download (default NULL).}

\item{dfp}{Target local directory for downloaded files (default "downloads").}

\item{url}{Server URL containing assay data.}

\item{show.files}{Whether to print server file data to console (default FALSE).}

\item{download}{Whether to download (TRUE) or return queried filename (FALSE).}

\item{tryload}{Whether to try loading downloaded data (default TRUE).}

\item{verbose}{Whether to return verbose messages.}

\item{sslver}{Whether to use server certificate check (default FALSE).}
}
\value{
New filepath to dir with downloaded data.
}
\description{
Uses RCurl to recursively download latest H5SE and HDF5 data objects the from server.
}
\examples{
path <- get_rmdl(which.class = "test", which.type = "h5se", tryload = FALSE)
unlink("downloads", recursive = TRUE)
}
