% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_db.R
\name{getdb}
\alias{getdb}
\alias{getdb_h5se_test}
\alias{getdb_h5_test}
\alias{getdb_h5se_gr}
\alias{getdb_h5se_gm}
\alias{getdb_h5se_rg}
\alias{getdb_h5_rg}
\title{Access database files.}
\usage{
getdb_h5se_test(
  namematch = "remethdb-h5se_gr-test.*",
  dfp = NULL,
  verbose = FALSE
)

getdb_h5_test(
  namematch = "remethdb-h5_rg-test_.*",
  dfp = NULL,
  verbose = FALSE
)

getdb_h5se_gr(name = "remethdb-h5se_gr_.*", dfp = NULL, verbose = FALSE)

getdb_h5se_gm(namematch = "remethdb-h5se_gm_.*", dfp = NULL, verbose = FALSE)

getdb_h5se_rg(namematch = "remethdb-h5se_rg_.*", dfp = NULL, verbose = FALSE)

getdb_h5_rg(namematch = "remethdb-h5_rg_.*", dfp = NULL, verbose = FALSE)
}
\arguments{
\item{namematch}{Filename pattern to match when searching for database 
(see defaults).}

\item{dfp}{Folder to search for database file 
(optional, if NULL then searches cache dir specified by BiocFileCache).}

\item{verbose}{Whether to return verbose messages (default FALSE).}
}
\value{
Either a SummarizedExperiment object for h5se functions, or a file 
path for h5 functions.
}
\description{
Combines download and load functions for databases. 
If the "namematch" argument isn't provided, the latest available file is downloaded.
All files include metadata for the available samples.

There are 6 functions. Functions with "h5se" access 
HDF5-SummarizedExperiment files, and "h5" functions access HDF5 databases. 
The 4 h5se functions are "rg" (RGChannelSet), "gm" (MethylSet), "gr" 
(GenomicRatioSet), and "test" (data for 2 samples from "gr"). The 2 h5 
functions are "rg" (red and green signal datasets), and "test" (data for 2 
samples from "rg"). See vignette for details about file types and classes.
}
\examples{
# download test file to temp directory
h5 <- getdb_h5_test(dfp = tempdir())
}
\seealso{
get_rmdl()
}
