% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_db.R
\name{getdb}
\alias{getdb}
\alias{getdb_h5se_test}
\alias{getdb_h5_test}
\alias{getdb_h5se_gr}
\alias{getdb_h5se_gm}
\alias{getdb_h5se_rg}
\alias{getdb_h5_rg}
\title{Access database files.}
\usage{
getdb_h5se_test(name = NULL, dfp = "downloads", verbose = FALSE)

getdb_h5_test(name = NULL, dfp = "downloads", verbose = FALSE)

getdb_h5se_gr(name = NULL, dfp = "downloads", verbose = FALSE)

getdb_h5se_gm(name = NULL, dfp = "downloads", verbose = FALSE)

getdb_h5se_rg(name = NULL, dfp = "downloads", verbose = FALSE)

getdb_h5_rg(name = NULL, dfp = "downloads", verbose = FALSE)
}
\arguments{
\item{name}{Database file name (optional, default NULL).}

\item{dfp}{Folder to search for database file specified by "name" 
(optional, default "downloads").}

\item{verbose}{Whether to return verbose messages (default FALSE).}
}
\value{
Either a SummarizedExperiment object for h5se functions, or a file path
for h5 functions.
}
\description{
Combines download and load functions for databases. 
If "name" argument not provided, the latest available file is downloaded.
All files include metadata for the available samples.

There are 6 functions. Functions with "h5se" access HDF5-SummarizedExperiment 
files, and "h5" functions access HDF5 databases. The 4 h5se functions are 
"rg" (RGChannelSet), "gm" (MethylSet), "gr" (GenomicRatioSet), and "test" 
(data for 2 samples from "gr"). The 2 h5 functions are "rg" (red and green 
signal datasets), and "test" (data for 2 samples from "rg"). See vignette 
for details about file types and classes.
}
\examples{
path <- getdb_h5se_test()
base::unlink("downloads", recursive = TRUE)
}
\seealso{
dldb(), get_rmdl()
}
