% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getrg}
\alias{getrg}
\title{Get raw signal data as either a list of `data.frame`'s or an `RGChannelSet` object}
\usage{
getrg(
  gsmv = "random",
  cgv = "all",
  dbn = "remethdb.h5",
  dat.type = c("se", "df"),
  dsv = c("redsignal", "greensignal"),
  metadata = TRUE
)
}
\arguments{
\item{gsmv}{Vector valid GSM IDs (rows) to query in the raw signal datasets.}

\item{cgv}{Vector of valid CpG probe addresses (columns) to query in the raw signal datasets.}

\item{dbn}{Name of the HDF5 database file.}

\item{dsv}{Vector of raw signal datasets or group paths to query, including both the red channel 'redsignal' and green channel 'greensignal' datasets.}

\item{metadata}{Whether to access available postprocessed metadata for queries samples.}

\item{data.type}{Format for returned query matches, either as datasets 'df' or `RGChannelSet` 'se' object.}
}
\value{
Returns either an `RGChannelSet` or list of `data.frame` objects from dataset query matches.
}
\description{
Retrieves query matches from raw signal HDF5 datasets. Handles identity queries to rows (GSM IDs) or columns (CpG probe addresses). Returns query matches either as a list of 2 `data.frame`s or a signle `RGChannelSet` object.
}
