% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{hread}
\alias{hread}
\title{Query and store an HDF5 dataset on row and column indices.}
\usage{
hread(ri, ci, dsn = "redsignal", dbn = "remethdb2.h5")
}
\arguments{
\item{ri}{rows indices in dataset.}

\item{ci}{columns indices in dataset.}

\item{dsn}{Name of dataset or group of dataset to connect with.}

\item{dbn}{Path to h5 database file.}
}
\value{
HDF5 database connection object.
}
\description{
Get a dataset connection object from an HDF5 database 
('.h5') file and return the indexed table subset.
}
\examples{
# Get tests data pointer
h5test.name <- "remethdb-h5_rg-test_0-0-1_1590090412.h5"
path <- system.file("extdata", h5test.name, package = "recountmethylation")
fn <- list.files(path)
dbpath = paste0(path, "/", fn)
# red signal, first 2 assay addr, 3 samples
reds <- hread(1:2, 1:3, d = "redsignal", dbn = dbpath)
dim(reds) # [1] 2 3
}
