% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{hread}
\alias{hread}
\title{Connect to an HDF5 dataset file.}
\usage{
hread(ri, ci, dsn = "redsignal", dbn = "remethdb.h5")
}
\arguments{
\item{ri}{rows indices in dataset.}

\item{ci}{columns indices in dataset.}

\item{dsn}{Name of dataset or group of dataset to connect with.}

\item{hfile}{Connection to an HDF5 database file.}
}
\value{
HDF5 database connection object.
}
\description{
Get a dataset connection object from an HDF5 database ('.h5') file.
}
