% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{hread}
\alias{hread}
\title{Query and store an HDF5 dataset on row and column indices.}
\usage{
hread(ri, ci, dsn = "redsignal", dbn = "remethdb2.h5")
}
\arguments{
\item{ri}{rows indices in dataset.}

\item{ci}{columns indices in dataset.}

\item{dsn}{Name of dataset or group of dataset to connect with.}

\item{dbn}{Path to h5 database file.}
}
\value{
HDF5 database connection object.
}
\description{
Get a dataset connection object from an HDF5 database ('.h5') file and return the indexed table subset.
}
\examples{
# get red signal for first 2 probe addresses, first 3 samples
st <- hread(1:3, 1:2, d = "redsignal", dbn = "remethdb2.h5")
}
\seealso{
data_mdpost
}
