% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnam_search_index.R
\name{query_si}
\alias{query_si}
\title{query_si}
\usage{
query_si(
  sample_idv,
  fh_csv_fpath,
  si_fname = "new_search_index",
  si_fpath = ".",
  lkval = c(1, 2)
)
}
\arguments{
\item{sample_idv}{Vector of valid sample IDs, or GSM IDs, which are included
in the rownames of the hashed features table at fh_csv_fpath (requried, 
vector of char strings).}

\item{fh_csv_fpath}{Path to the hashed features table, which includes rownames
corresponding to sample ID strings in the sample_idv vector (required, char).}

\item{si_fname}{Base filename of the search index object, used to find the 
search index and index dict files, which are expected to be located at
si_fapth (required, char).}

\item{si_fpath}{Path to the directory containing the search index and index
dict files (required, char).}

\item{lkval}{Vector of K nearest neighbors to return per query (optional, 
int, c(1,2)).}
}
\value{

}
\description{
Query an HNSW search index. Does K Nearest Neighbors lookup on a previously 
saved search index object, returning the K nearest neighbors of the queried 
sample(s). The `query_si()` function returns verbose output, which can be 
silenced with suppressMessages()`.
}
