% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{rgse}
\alias{rgse}
\title{Form an `RGChannelSet` from signal data}
\usage{
rgse(ldat, verbose = FALSE)
}
\arguments{
\item{ldat}{List of raw signal data query results. Must include 2 
`data.frame` objects named 'redsignal' and 'greensignal'.}

\item{verbose}{Whether to post status messages.}
}
\value{
Returns a `RGChannelSet` object from raw signal dataset queries.
}
\description{
Forms an object of `RGChannelSet` class from a signal data object 
returned from a query to the red and green signal tables.
}
\examples{
path <- system.file("extdata", h5test.name, package = "recountmethylation")
fn <- list.files(path)
dbpath = paste(path, fn, sep = "/")
rg = getrg(dbn = dbpath, all.gsm = TRUE, metadata = FALSE)
dim(rg) # [1] 11162     2
class(rg)
# [1] "RGChannelSet"
# attr(,"package")
# [1] "minfi"
}
\seealso{
getrg()
}
