% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{rgse}
\alias{rgse}
\title{Form an `RGChannelSet` from signal data}
\usage{
rgse(ldat, verbose = FALSE)
}
\arguments{
\item{ldat}{List of raw signal data query results. Must include 2 
`data.frame` objects named 'redsignal' and 'greensignal'.}

\item{verbose}{Whether to post status messages.}
}
\value{
Returns a `RGChannelSet` object from raw signal dataset queries.
}
\description{
Forms an object of `RGChannelSet` class from a signal data object 
returned from a query to the red and green signal tables.
}
\examples{
# get the list of datasets for all probe addresses, 3 samples
path <- system.file("extdata", "testh5", package = "recountmethylation")
dbpath <- paste0(path, "/remethdbtest.h5")
gsml = c("GSM1235984", "GSM1236090", "GSM1506278")
ldat = getrg(gsmv = gsml, dbn = dbpath, data.type = "df", metadata = FALSE)
rg = rgse(ldat) # get the rg set object
class(rg)
dim(minfi::getBeta(rg))
}
\seealso{
getrg()
}
