% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{rgse}
\alias{rgse}
\title{Form an object of class `RGChannelSet` from a raw signal dataset query}
\usage{
rgse(ldat, verbose = FALSE)
}
\arguments{
\item{ldat}{List of raw signal data query results. Must include 2 `data.frame` objects named 'redsignal' and 'greensignal'.}

\item{verbose}{Whether to post status messages.}
}
\value{
Returns a `RGChannelSet` object from raw signal dataset queries.
}
\description{
Forms a `RGChannelSet` object from an HDF5 databse file query to the red and green raw signal tables. See `getrg()` function for implementation.
}
