% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{rgse}
\alias{rgse}
\title{Form an `RGChannelSet` from a signal table query}
\usage{
rgse(ldat, verbose = FALSE)
}
\arguments{
\item{ldat}{List of raw signal data query results. Must include 2 `data.frame` objects named 'redsignal' and 'greensignal'.}

\item{verbose}{Whether to post status messages.}
}
\value{
Returns a `RGChannelSet` object from raw signal dataset queries.
}
\description{
Forms an object of `RGChannelSet` class from a query to the red and green signal tables.
}
\examples{
# get the list of datasets for all probe addresses, 3 samples
ldat = getrg(gsmv = c("GSM1235984", "GSM1236090", "GSM1506278"), data.type = "df", metadata = F)

# get the rg set object
rg = rgse(ldat)
}
\seealso{
getrg()
}
