% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{servermatrix}
\alias{servermatrix}
\title{servermatrix}
\usage{
servermatrix(
  dn,
  sslver = FALSE,
  url = "https://recount.bio/data/",
  printmatrix = TRUE,
  verbose = FALSE,
  recursive = TRUE
)
}
\arguments{
\item{dn}{Server data returned from RCurl.}

\item{sslver}{Whether to use SSL certificate authentication for server connection (default FALSE).}

\item{printmatrix}{Whether to print the data matrix to console (default TRUE).}

\item{verbose}{Whether to show verbose messages (default FALSE).}

\item{recursive}{Whether to recursively grab file sizes for h5se object (default TRUE).}
}
\value{
dm matrix of server files and file metadata
}
\description{
Get a matrix of database files from the recount server. Called by get_rmdl.
}
\examples{
dn <- RCurl::getURL("https://recount.bio/data/", .opts = list(ssl.verifypeer = FALSE))
sm <- servermatrix(dn)
}
\seealso{
get_rmdl
}
